Update install.py
Replaced some of the larger functions with a newly created Import function which dynamicly fetches needed modules in order for operation to work. These modules can be downloaded manually ofc and placed in the ./ folder of this script and the module.py part will look locally first and hand over the requested module.